RateLimiter.jl Documentation

RateLimiter.jl Documentation

Functions

Index

Functions

RateLimiter.AbstractRateLimiter — Type

Abstract base type for all rate limiters.

RateLimiter.NoLimitRateLimiter — Type

A rate limiter which does not limit the rate.

RateLimiter.TokenBucketRateLimiter — Type

Token-Bucket rate limiter. This limiter is used to control for constraints such as bandwidth and burstiness.

See: https://en.wikipedia.org/wiki/Token_bucket#Algorithm

RateLimiter.@rate_limit — Macro

@rate_limit(limiter, cost, expr)

A macro to limit the rate an expression can be called.

Arguments

- `limiter`: rate limiter
- `cost`: cost of executing the method.
- `expr`: expression to be evaluated

Index

RateLimiter.AbstractRateLimiter

RateLimiter.NoLimitRateLimiter

RateLimiter.TokenBucketRateLimiter

RateLimiter.@rate_limit